feat: TASK-009 テストファイル整理・coverage.exclude 修正
- sessions/queue.test.ts, registry.test.ts, connectors/slack/threads.test.ts を
  __tests__/ サブディレクトリに移動（AC-E003-01）
- node:test スタイルから vitest スタイル（describe/it/expect）に統一
- AC-ID（AC-E003-01）を各 describe ブロックに付与
- coverage.exclude の "src/**/*.test.ts" は既に設定済みで AC-E003-02 充足確認
- pnpm test: 228 tests PASS / pnpm test --coverage: branch 9.96% 正常計測

Closes #53
